nature app 분리완료
