+++ config/urls.py
@@ -18,8 +18,9 @@ from django.urls import path, include
 from pybo.views import base_views
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
     path('pybo/', include('pybo.urls')),
     path('common/', include('common.urls')),
+    path('nature/', include('nature.urls')),
+    path('admin/', admin.site.urls),
     path('', base_views.index, name='index'),  # '/' 에 해당되는 path
 ]
