+++ nature/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class NatureConfig(AppConfig):
+    name = 'nature'
