+++ templates/navbar.html
@@ -13,10 +13,10 @@
     			{% else %}
 		    	<a class="nav-link" href="{% url 'common:login' %}">로그인</a>
 			{% endif %}
-		<li class="nav-item "><a class="nav-link" href="{% url 'common:forest' %}">서울의 자연관광지</a></li>
-		<li class="nav-item "><a class="nav-link" href="{% url 'common:forest2' %}">경기도의 자연관광지</a></li>
-		<li class="nav-item "><a class="nav-link" href="{% url 'common:forest4' %}">강원도의 자연관광지</a></li>
-		<li class="nav-item "><a class="nav-link" href="{% url 'common:forest3' %}">제주도의 자연관광지</a></li>
+		<li class="nav-item "><a class="nav-link" href="{% url 'nature:forest' %}">서울의 자연관광지</a></li>
+		<li class="nav-item "><a class="nav-link" href="{% url 'nature:forest2' %}">경기도의 자연관광지</a></li>
+		<li class="nav-item "><a class="nav-link" href="{% url 'nature:forest4' %}">강원도의 자연관광지</a></li>
+		<li class="nav-item "><a class="nav-link" href="{% url 'nature:forest3' %}">제주도의 자연관광지</a></li>
 		</li>
 	    </ul>
 	</div>
